feat(hooks): error if `useEffectListener` isn’t passed a dependency list
Not passing a dependency list can result in infinite loops, so was 
always an error and not permitted in Typescript. This change ensures 
consumers get clear error feedback if they don’t include the dependency 
list.

BREAKING CHANGE: If previously `useEffectListener` was used without 
passing a dependency list (which may have worked in some edge cases), 
this will now cause `useEffectListener` to throw an error.