feat: Rewrite client-side effects emitter & add plain JS support (#367)
Closes #148

This commit extracts all the logic responsible for processing boardgame.io state updates and emitting effects from the React component to a plain JavaScript class.

Doing so helps clarify some logic, makes it possible to use the effects emitter with a plain JS boardgame.io client, and also looks like it has some small performance benefits at least in terms of a reduced number of React renders (based on some simple profiling with the existing demo).

BREAKING CHANGE: The public API for the existing packages has not changed and should continue to work as previously. The React unit tests remain unchanged. However, because this is such a thorough rewrite, there may be small differences that turn out to be breaking in subtle edge cases.

We are also now using an internal type that was added in boardgame.io@0.42.0, so no longer support 0.39.16, ^0.40 and ^0.41.