feat(emitter): Use `setTimeout` if `requestAnimationFrame` doesn’t run
Partially addresses #57. `requestAnimationFrame` is usually paused when a browser tab is inactive or in the background. We now schedule a `setTimeout` each time we request a new frame, which will only get run if the rAF callback is not called within 1000ms. `setTimeout` is also usually throttled, but will more likely run in the background than `requestAnimationFrame`.